update style on project detail
