+++ public/index.html
@@ -4,9 +4,10 @@
     <meta charset="utf-8" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
     <link
-      href="https://fonts.googleapis.com/css?family=Fira+Sans&display=swap"
+      href="https://fonts.googleapis.com/css?family=Playfair+Display:700&display=swap"
       rel="stylesheet"
     />
+
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <!--
